fix(recompute): tolerate profile and template kinds in coalesced recompute
An updated profile or template node in a merge or rebase change set aborted the
whole coalesced recompute. Resolving its changed relationship fields went through
get_node, which rejects anything that is not a NodeSchema, and the raise was
swallowed by the coordinator, leaving every derived value the merge touched stale.

Resolve relationships through has/get instead. They accept any node-like kind
(profiles and templates carry relationships but are not NodeSchema) and an absent
kind, so one such node in the change set no longer discards the recompute for the rest.
